requirements.txt
# Core dependencies
Pillow>=10.0.0

# API dependencies (optional)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Development dependencies (optional)
black>=23.0.0
pytest>=7.4.0
